#127 - added redirect after social login failure

+++ client/src/app/pages/social-login-page/social-login-page.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { AuthenticationService } from "../../services/authentication.service";
-import { Router } from "@angular/router";
+import { ActivatedRoute, Router } from "@angular/router";
 import { CookieService } from 'ngx-cookie-service';
 
 @Component({
@@ -9,21 +9,32 @@ import { CookieService } from 'ngx-cookie-service';
   styleUrls: ['./social-login-page.component.css']
 })
 export class SocialLoginPageComponent implements OnInit {
+  success: boolean;
+  error: boolean;
 
   constructor(private authenticationService: AuthenticationService,
               private cookieService: CookieService,
-              private router: Router) {
+              private router: Router,
+              private route: ActivatedRoute,) {
   }
 
   ngOnInit() {
+    this.route.queryParams.subscribe((queryParams) => {
+      this.success = queryParams['success'];
+    });
+
+    console.log(this.success);
+
     const token = this.cookieService.get('social-authentication');
-    if (token.length) {
+    if (token.length && this.success) {
       this.authenticationService.loginWithToken(token, false).then(() => {
         this.cookieService.delete('social-authentication');
         this.router.navigate(['']);
       }, () => {
         this.router.navigate(['login'], { queryParams: { 'socialLoginSuccess': 'false' } });
       });
+    } else {
+      this.router.navigate(['login'], { queryParams: { 'socialLoginSuccess': 'false' } });
     }
   }
 }
